+++ gui/__init__.py
@@ -1,0 +1,4 @@
+from .main import DownloaderGUI
+from .config_dialog import ConfigDialog
+
+__all__ = ["DownloaderGUI", "ConfigDialog"]
